requirements.txt
mmdet
lyft_dataset_sdk
py_trees
absl-py
addict
aiohttp
aiosignal
antlr4-python3-runtime
async-timeout
asynctest
attrs
backcall
beartype
black
bleach
cachetools
certifi
chardet
charset-normalizer
click
colorama
commonmark
configparser
cycler
decorator
defusedxml
descartes
dictor
diskcache
distlib
docker-pycreds
einops
elementpath
entrypoints
ephem
exceptiongroup
filelock
filterpy
fire
flake8
fonttools
frozenlist
fsspec
future
gitdb
GitPython
google-auth
google-auth-oauthlib
gql
graphql-core
greenlet
grpcio
huggingface-hub
hydra-core
idna
imageio
imageio-ffmpeg
imgaug
importlib-metadata
importlib-resources
iniconfig
ipdb
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-console
jupyter_client
jupyter_core
kiwisolver
llvmlite
lyft-dataset-sdk
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mistune
mkl-fft
mkl-random
mkl-service==2.4.0
mmcls==0.24.1
mmsegmentation==0.29.1
model-index==0.1.11
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.6
networkx==2.2
notebook==6.0.3
numba==0.53.0
numpy
nuscenes-devkit
oauthlib
olefile
omegaconf==2.2.3
opencv-python==4.2.0.32
openmim==0.3.3
ordered-set==4.1.0
packaging==21.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.10.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.5.4
plotly==5.11.0
pluggy==1.0.0
plyfile==0.7.4
prettytable==3.5.0
proglog==0.1.10
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.20.3
psutil==5.7.0
ptyprocess==0.6.0
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.9.1
pyDeprecate==0.3.1
pydot==1.4.1
pyflakes==2.5.0
pygame==2.0.1
Pygments==2.6.1
pynvim==0.4.3
pyparsing==2.4.6
pyquaternion==0.9.9
pyrsistent==0.16.0
pytest==7.2.0
python-dateutil==2.8.2
pytictoc==1.5.2
pytorch-lightning==1.5.10
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
ray==2.1.0
rdp==0.8
regex==2022.10.31
requests==2.23.0
requests-oauthlib==1.3.0
rich==12.6.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
sentry-sdk==0.14.2
Shapely==1.7.0
shortuuid==1.0.1
six
smmap==3.0.1
subprocess32==3.5.4
tabulate==0.8.7
tenacity==8.1.0
tensorboard==2.11.0
tensorboard
termcolor==2.1.0
terminado==0.8.3
terminaltables==3.1.10
testpath==0.4.4
timm==0.5.4
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.43.0
traitlets==4.3.3
transformers==4.24.0
trimesh==2.35.39
typed-ast==1.5.4
typing_extensions==4.4.0
ujson==5.3.0
urllib3==1.25.8
wandb==0.8.29
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==3.5.1
xmlschema==1.0.18
yapf==0.32.0
yarl==1.8.1
zipp==3.1.0
